Plex and plexamp plugin unit tests
A development pattern I used repeatly while testing and updating parts of the plugin
runs as follows:
   1) develop code
   2) sync code to a raspberry pi
   3) volumio plugin package
   4) create and move to package install directory
   5) unzip plugin zip file
   6) volumio plugin install
   7) restart the raspberry pi
   8) find problems with the plugin
   9) instrument plugin code to localize problems
  10) repeat steps 7-9 multiple times to root cause the problem
  11) uninstall the plugin and go to step 1

This is a cumbersome process that takes diagnosis out of developement environment into
a raspberry pi runtime environment, where tools are simpler, fewer, and farther between.

This points toward the need for more testing tools in the development environment where
code correctness can be verified before it is deployed to a runtime environment.

This change starts with a series of unit tests written for the jest framework that
test parts of the plugin settings process, leading up to a several tests of
getPinAndUpdateConfig.  The tests are located in getpinupdateandconfig.test.js in the
/tests directory.

The bulk of the testing effort was to create a series of mock web services that
simulate the behavior of PlexTV and two media servers.  These are implemented using
the mock-http-server npm module. Much of the configuration consists of response
data, either in the form of XML or JSON data to be returned in the body of HTTP
responses, or as dictionaries configuring different HTTP response messages themselves.
The dictionaries are grouped by HTTP request type: head_resp for HEAD queries,
get_resp for GET, and post_resp for POST.  Each entry is a function that takes a single
arugment -- a v-conf style configuration struct with get/set attribute semantics.
The data returned by each function configurues a particular endpoint for mock-http-server.

So, the PlexTV server configuration necessary to run getPinAndUpdateConfig consists of
five mock endopints that can be inserted into a test like this:

            server
                .on(mockData.head_resp['401_success_xml_headers_no_body'](config))
                .on(mockData.get_resp['eval_auth_root'](config))
                .on(mockData.get_resp['eval_auth_pms_local_resources'](config))
                .on(mockData.post_resp['pins_xml'](config))
                .on(mockData.get_resp['pin_plex_token'](config));

If we want to replace the response for a particular service request to introduce errors,
(with say, the /pms/resources request to get a media server list), we only need to create
a new dictionary entry configuring the response:

    'pms_local_resources_500_failure': function(config) {
        return {
            method: "GET",
            path: "/pms/resources",
            reply: {
                status: 500,
                headers: html_headers,
                body: "<html><h1>Server Failure</h1></html>"
            }
        };
    },

and then update the test server configuration:

            server
                .on(mockData.head_resp['401_success_xml_headers_no_body'](config))
                .on(mockData.get_resp['eval_auth_root'](config))
                .on(mockData.get_resp['pms_local_resources_500_failure'](config))
                .on(mockData.post_resp['pins_xml'](config))
                .on(mockData.get_resp['pin_plex_token'](config));

This should now cause plexcloud.getServers() to fail and unit test code should
verify responses accordingly.

To enable more testing, a timeout option is added to PlexPinAuth requests for getToken and getPin.
Such a timeout is also useful for plugin operation, because timeouts otherwise default to the
timeout behavior of TCP connection behavior for version of Linux that volumio is running on.

Testing revealed a number code problems:

- Error message typing was moved from detectAndSavePlexToken to getPinAndUpdateConfig, where
it eliminated a level of condition checking.

- There were a number syntax errors (e.g. using variables named error vs err, missing calls
to defer.reject, missing .fail actions to call defer.reject) that prevented errors from
being propagated correctly.

- localServer2 was not being configured correctly in some tests.