getuiconfig() unit tests and a few small improvements
Added the capability for CommandCoreMock to save JSON data to an internal struct for later
inspeciton.  This was discovered to not be needed to test getuiconifig() because the routine
returns the modified data.  I left it in place because it may be useful in future.

Changed the names of a few variables in CommandCoreMock to make them more readable.

Added the executeOnPlugin() method to CommandCoreMock because it was required for getIP();

Running the new tests revealed that Jest runs different test groups in parallel.  For tests
that set up web service mocks like those in getuiconfig.test.js and getpinandupdateconfig.test.js,
this means that those tests must run against different ports.  This required that the mock
Config() object now include port number information for the two "local" plex media server
mocks.  The Plex cloud mock responses to identify those ports are now defined with the
symbols LOCAL_PORT_1 and LOCAL_PORT_2.  These are now dynamically replaced when the local
media server mock endpoints are called using the eval_resource_replace() call.

Tests in getpinandupdateconfig.test.js were originally named simply "plex.js".  Since that's
wrong and not very descriptive, it is now changed to 'index.getpinandupdateconfig.js tests'.

Four additional tests of getuiconfig() are now located in getuiconfig.test.js.