Plex search improvements
User testing with the search box of the plex plugin showed that something was improperly crashing --
a status bar crept slowly accross the top of the interface, and never completed.  This turned out to
be caused by the fact that the code in ControllerPlexAmp.prototype.search was experiencing two problems:

1) The query value was being passed as raw text rather than as a URI-encoded compenent, causing the
underlying HTTP queries to fail when characters like spaces between words appeared, and
2) The query responses were being treated as if searches for artists, albums, and songs were always
successful and passed to member functions for formatting even no responses matched.

The second problem occurs because of how plex.prototype.query() works:

                                if (result.MediaContainer.Metadata) {
                                    cacheSet(cacheKey, result.MediaContainer.Metadata);
                                    defer.resolve(result.MediaContainer.Metadata);
                                } else {
                                    cacheSet(cacheKey, result.MediaContainer);
                                    defer.resolve(result.MediaContainer);
                                }

If the search has a keyword match, the MediaContainer response includes an arrray of Metadata. This
metatdata is returned, and iterated upon for formatting.   If there is no match, the MediaContainer
object itself is returned.  This is not an array, and attempts to iterate over it crash.

The first problem was fixed just by simply adding and appropriate encodeURIComponent() call.

The second was fixed by looking to see if the result has a "size" property and that size is 0.  When
true, this indicates wthat it is the MediaContainer object with no attached Metadata, and formatting
of the result is skipped.

Once the mechanics of search worked, an additional problem was discovered.  Returned data for objects
and albums did not create links on the interface to the approprate artist and album pages.  This turned
out to be because the ControllerPlexAmp.prototype.search function was formatting the URL compoents
passed to handleBrowserUri() with the prefixes "plexamp/artist" and "plexamp/album" for artists and
albums respectively.  Neither of these uri components is recognized by handleBrowserUri().

This was fixed by changing the prefixes "plexamp/artists" and "plexamp/allalbums", which already provide
the correct behaviors.